+++ sdk/platform/Content/ContentPlatformClient.js
@@ -1357,11 +1357,15 @@ class Content {
    * @description: Use this API to retrieve the custom objects. - Check out [method documentation](https://partners.fynd.com/help/docs/sdk/platform/content/getCustomObjects/).
    */
   async getCustomObjects(
-    { requestHeaders } = { requestHeaders: {} },
+    { pageNo, pageSize, definitionId, requestHeaders } = { requestHeaders: {} },
     { responseHeaders } = { responseHeaders: false }
   ) {
     const { error } = ContentPlatformValidator.getCustomObjects().validate(
-      {},
+      {
+        pageNo,
+        pageSize,
+        definitionId,
+      },
       { abortEarly: false, allowUnknown: true }
     );
     if (error) {
@@ -1372,7 +1376,11 @@ class Content {
     const {
       error: warrning,
     } = ContentPlatformValidator.getCustomObjects().validate(
-      {},
+      {
+        pageNo,
+        pageSize,
+        definitionId,
+      },
       { abortEarly: false, allowUnknown: false }
     );
     if (warrning) {
@@ -1383,6 +1391,9 @@ class Content {
     }
 
     const query_params = {};
+    query_params["definition_id"] = definitionId;
+    query_params["page_no"] = pageNo;
+    query_params["page_size"] = pageSize;
 
     const xHeaders = {};