+++ sdk/platform/FileStorage/FileStoragePlatformModel.d.ts
@@ -264,10 +264,6 @@ export = FileStoragePlatformModel;
  * @property {UpiQrcodeGenerator} [upi_qrcode_generator]
  */
 /**
- * @typedef Params
- * @property {string} [subpath] - The subpath for the file.
- */
-/**
  * @typedef PaymentData
  * @property {number} [amount]
  * @property {string} [date]
@@ -446,7 +442,7 @@ export = FileStoragePlatformModel;
  * @typedef StartRequest
  * @property {string} content_type
  * @property {string} file_name
- * @property {Params} [params]
+ * @property {Object} [params]
  * @property {number} size
  * @property {string[]} [tags]
  */
@@ -519,7 +515,7 @@ export = FileStoragePlatformModel;
 declare class FileStoragePlatformModel {
 }
 declare namespace FileStoragePlatformModel {
-    export { AwbNumberLabelBarcodeGenerator, Brand, CDN, Cgst, CompanyDetail, CompleteResponse, ConversionRate, CopyFiles, CreatedBy, CustomerBillingDetail, CustomerShippingDetail, DeliveryPartnerDetail, DestinationNamespace, DigitalsignatureGenerator, Document, DummyTemplateData, DummyTemplateDataItems, DummyTemplateDataPayload, ExtensionSlug, FailedResponse, Igst, Image, Inr, InvoiceDetail, InvoiceTypesDataResponse, InvoiceTypesResponse, ItemsProductTable, Kwargs, KwargsAwbNumber, KwargsUpiQrcode, Meta, MetaProperty, Params, PaymentData, PaymentReceiptCustomerDetails, PaymentReceiptFormat, PaymentReceiptMeta, PaymentReceiptOrderDetails, PaymentReceiptPayload, PaymentReceiptPayments, PaymentReceiptRequestBody, PaymentReceiptService, PaymentReceiptTaxes, PdfConfig, PdfConfigSaveSuccess, PdfConfigSaveSuccessData, PdfConfigSuccess, PdfConfigSuccessData, PdfDefaultTemplateSuccess, ProductTable, Rates, RegisteredCompanyDetail, ReturnDetail, Sgst, ShipmentIdBarcodeGenerator, SignedQrcodeGenerator, SignUrlRequest, SignUrlResponse, StartRequest, StartResponse, StoreDetail, Tax, Taxes, TaxTable, UpiQrcodeGenerator, Upload, Urls, Usd };
+    export { AwbNumberLabelBarcodeGenerator, Brand, CDN, Cgst, CompanyDetail, CompleteResponse, ConversionRate, CopyFiles, CreatedBy, CustomerBillingDetail, CustomerShippingDetail, DeliveryPartnerDetail, DestinationNamespace, DigitalsignatureGenerator, Document, DummyTemplateData, DummyTemplateDataItems, DummyTemplateDataPayload, ExtensionSlug, FailedResponse, Igst, Image, Inr, InvoiceDetail, InvoiceTypesDataResponse, InvoiceTypesResponse, ItemsProductTable, Kwargs, KwargsAwbNumber, KwargsUpiQrcode, Meta, MetaProperty, PaymentData, PaymentReceiptCustomerDetails, PaymentReceiptFormat, PaymentReceiptMeta, PaymentReceiptOrderDetails, PaymentReceiptPayload, PaymentReceiptPayments, PaymentReceiptRequestBody, PaymentReceiptService, PaymentReceiptTaxes, PdfConfig, PdfConfigSaveSuccess, PdfConfigSaveSuccessData, PdfConfigSuccess, PdfConfigSuccessData, PdfDefaultTemplateSuccess, ProductTable, Rates, RegisteredCompanyDetail, ReturnDetail, Sgst, ShipmentIdBarcodeGenerator, SignedQrcodeGenerator, SignUrlRequest, SignUrlResponse, StartRequest, StartResponse, StoreDetail, Tax, Taxes, TaxTable, UpiQrcodeGenerator, Upload, Urls, Usd };
 }
 /** @returns {AwbNumberLabelBarcodeGenerator} */
 declare function AwbNumberLabelBarcodeGenerator(): AwbNumberLabelBarcodeGenerator;
@@ -817,14 +813,6 @@ type MetaProperty = {
     signed_qrcode_generator?: SignedQrcodeGenerator;
     upi_qrcode_generator?: UpiQrcodeGenerator;
 };
-/** @returns {Params} */
-declare function Params(): Params;
-type Params = {
-    /**
-     * - The subpath for the file.
-     */
-    subpath?: string;
-};
 /** @returns {PaymentData} */
 declare function PaymentData(): PaymentData;
 type PaymentData = {
@@ -1039,7 +1027,7 @@ declare function StartRequest(): StartRequest;
 type StartRequest = {
     content_type: string;
     file_name: string;
-    params?: Params;
+    params?: any;
     size: number;
     tags?: string[];
 };